LFS-106: Custom parser for NCIT flat. Add VocabularyIndexException for handling errors and split flat parsing from servlet into own class.
